requirements.txt
# ─── Web Framework ────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ─── CLI ──────────────────────────────────────────────────────────
questionary>=2.0.1       # arrow-key menus
rich>=13.7.0             # panels, tables, progress bars

# ─── Task Scheduler ───────────────────────────────────────────────
apscheduler>=3.10.4

# ─── Configuration ────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ─── LLM Providers ────────────────────────────────────────────────
openai>=1.30.0          # OpenAI + Ollama/LM Studio (OpenAI-compatible)
httpx>=0.27.0           # HTTP client for Ollama REST API

# ─── Vector Store ─────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Document Processing ──────────────────────────────────────────
python-docx>=1.1.2      # Read/write .docx files

# ─── Google Drive ─────────────────────────────────────────────────
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ─── Testing ──────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
